requirements.txt
_libgcc_mutex==0.1
_openmp_mutex==5.1
absl-py==2.3.1
addict==2.4.0
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
ca-certificates==2025.2.25
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
chumpy==0.70
click==8.1.8
colorama==0.4.6
contourpy==1.1.1
crcmod==1.7
cryptography==44.0.2
cycler==0.12.1
cython==3.0.12
easydict==1.13
einops==0.8.1
faiss-gpu==1.7.2
filelock==3.14.0
fonttools==4.56.0
fsspec==2025.3.0
ftfy==6.2.3
google-auth==2.40.3
google-auth-oauthlib==1.0.0
grpcio==1.70.0
huggingface-hub==0.29.3
idna==3.10
importlib-metadata==8.5.0
importlib-resources==6.4.5
jinja2==3.1.6
jmespath==0.10.0
joblib==1.4.2
json-tricks==3.17.3
kiwisolver==1.4.7
ld_impl_linux-64==2.40
libffi==3.4.4
libgcc-ng==11.2.0
libgomp==11.2.0
libstdcxx-ng==11.2.0
markdown==3.7
markdown-it-py==3.0.0
markupsafe==2.1.5
matplotlib==3.7.5
mdurl==0.1.2
metric-learn==0.7.0
mmcv==2.1.0
mmdet==3.3.0
mmengine==0.10.7
mmpose==1.3.2
mmsegmentation==1.2.2
model-index==0.1.11
mpmath==1.3.0
munkres==1.1.4
ncurses==6.4
networkx==3.1
nltk==3.9.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.1.105
oauthlib==3.3.1
opencv-python==4.11.0.86
opendatalab==0.0.10
openmim==0.3.9
openssl==3.0.16
openxlab==0.1.2
ordered-set==4.1.0
oss2==2.17.0
packaging==24.2
pandas==2.0.3
pillow==10.4.0
pip==24.2
platformdirs==4.3.6
pplr==1.0.0
prettytable==3.11.0
protobuf==5.29.5
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycocotools==2.0.7
pycparser==2.22
pycryptodome==3.22.0
pygments==2.19.1
pyparsing==3.1.4
python==3.8.20
python-dateutil==2.9.0.post0
pytz==2023.4
pyyaml==6.0.2
readline==8.2
regex==2024.11.6
requests==2.28.2
requests-oauthlib==2.0.0
rich==13.4.2
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.3.2
scipy==1.10.1
segment-anything==1.0
setuptools==60.2.0
shapely==2.0.7
six==1.17.0
sqlite==3.45.3
sympy==1.13.3
tabulate==0.9.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
terminaltables==3.1.10
threadpoolctl==3.5.0
timm==1.0.15
tk==8.6.14
tomli==2.2.1
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1
tqdm==4.65.2
triton==3.0.0
typing-extensions==4.12.2
tzdata==2025.2
urllib3==1.26.20
wcwidth==0.2.13
werkzeug==3.0.6
wheel==0.44.0
xtcocotools==1.14.3
xz==5.6.4
yacs==0.1.8
yapf==0.43.0
zipp==3.20.2
zlib==1.2.13
